Create a data.frame with estimated values from a SAMtool assessment method used in an MSE

Create a data.frame with estimated values from a SAMtool assessment method used in an MSE

Usage

get_Assess_Estimates(x, model = "Model 1")

# S3 method for class 'MSE'
get_Assess_Estimates(x, model = "Model 1")

# S3 method for class 'list'
get_Assess_Estimates(x, model = NULL)

# S3 method for class 'MMSE'
get_Assess_Estimates(x, model = NULL)

Arguments

x

An object of class MSE or a list of MSE objects, where MSE includes management procedures that use SAMtool stock assessment functions that return estimated values in MSE@PPD.

model

An optional name for the model. If x is a list of objects, model will be taken from names(x). If names(x) is NULL, model will be given sequential numerical values (e.g., Model 1, Model 2, ...)

Value

A data.frame with columns:

Year_assess	The year the assessment was run in the MSE

Year_est	The year corresponding with the estimated value

Variable	The estimated variable

Value	The estimated value

MP	The name of the management procedure

Simulation	The simulation number

Model	The name of model