Extract the meta-data from a Hist or MSE object

Extract the meta-data from a Hist or MSE object

Usage

get_Metadata(x)

# S3 method for class 'Hist'
get_Metadata(x)

# S3 method for class 'MSE'
get_Metadata(x)

# S3 method for class 'list'
get_Metadata(x)

# S3 method for class 'MMSE'
get_Metadata(x)

Arguments

x

An object of class Hist, MSE, or a list of Hist or MSE objects

Value

A named list with elements:

nsim	The number of simulations

nage	The number of age classes

Ages	The age classes

nyear	The number of historical years

Hist.Years	A data.frame with the historical years in the Year column

proyears	The number of projection years

Pro.Years	A data.frame with the projection years in the Year column

All.Years	A data.frame with the historical and the projection years in the Year column

nMPs	The number of MPs (if x is an object of class MSE)

MPs	The MPs (if x is an object of class MSE)

Details

If x is a list of objects, each object must have identical structure, i.e., same number of simulations, same number of age-classes, historical and projection years, management procedures, etc