Plot at-Age schedules

Plots Length, Weight, Maturity, Natural-Mortality, Selectivity, and Retention-at-Age schedules.

Usage

plot_at_Age(
  x,
  quantiles = c(0.025, 0.975),
  scale = NULL,
  variable = "Length",
  xlab = "Age (Year)",
  ylab = NULL,
  title = "",
  years = NULL,
  alpha = 0.1,
  lwd = 1,
  use_theme = NULL,
  colpalette = "Dark2",
  print = TRUE
)

plot_Length(x, ...)

plot_Weight(x, ...)

plot_Maturity(x, ...)

plot_N.Mortality(x, ...)

plot_Select(x, ...)

plot_Retention(x, ...)

plot_Select_Maturity(x, ...)

Arguments

x

An object of class Hist, MSE, or a list of Hist or MSE objects

quantiles

Lower and upper quantiles to calculate. Numeric vector of length 2.

scale

An optional function with a single numeric argument that returns transformed or scaled numeric values. See lb2kg and kg2lb for example. Can be a list of functions for list objects (NA for no transformation)

variable

String. One of 'Length', 'Weight', 'N.Mortality','Maturity', 'Select', 'Retention'

xlab

X-axis label (default 'Age (Year)')

ylab

Y-axes label

title

Optional title

years

Optional numeric vector specifying the years to plot. Default is the first and last historical year, and the last projection year

alpha

Transparency parameter

lwd

Line width

use_theme

Optional ggplot theme

colpalette

Color palette from RColorBrewer

print

Logical. Print the plot?

...

Named arguments passed to plot_at_Age

Value

A named list with:

p	The ggplot object

df	Data.frame with the summary statistics (median and quantiles)

Details

Note that the Selectivity and Retention curves in these plots are from the operating model. If an MP changes the selectivity/retention, this is not shown in these plots.

Functions

plot_Length(): Plot Length-at-Age

plot_Weight(): Plot Weight-at-Age

plot_Maturity(): Plot Maturity-at-Age

plot_N.Mortality(): Plot N.Mortality-at-Age

plot_Select(): Plot Selectivity-at-Age

plot_Retention(): Plot Retention -at-Age

plot_Select_Maturity(): Plot Selectivity-, Retention-, and Maturity-at-Age